Pc add_remove, list,view flow.
