+++ apps/front-end/src/pages/admin/pcScreen/List.js
@@ -35,6 +35,7 @@ import {
   getOptions,
   getSelectedProgramId,
   tableCustomStyles,
+  cohortService,
 } from "@shiksha/common-lib";
 // import Table from "./Table";
 import { useTranslation } from "react-i18next";
@@ -56,10 +57,6 @@ const uiSchema = {
   },
 
   block: {
-    "ui:widget": MultiCheck,
-    "ui:options": {},
-  },
-  status: {
     "ui:widget": MultiCheck,
     "ui:options": {},
   },
@@ -99,19 +96,6 @@ const schemat = {
       },
       uniqueItems: true,
     },
-    status: {
-      type: "array",
-      title: "STATUS",
-      grid: 1,
-      _hstack: {
-        maxH: 130,
-        overflowY: "scroll",
-      },
-      items: {
-        type: "string",
-      },
-      uniqueItems: true,
-    },
   },
 };
 export default function List({ footerLinks, userTokenInfo }) {
@@ -123,7 +107,6 @@ export default function List({ footerLinks, userTokenInfo }) {
   const [filter, setFilter] = useState({});
   const [loading, setLoading] = useState(true);
   const [tableLoading, setTableLoading] = useState(true);
-  const [modal, setModal] = useState(false);
   const [data, setData] = useState([]);
   const [paginationTotalRows, setPaginationTotalRows] = useState(0);
   const [enumOptions, setEnumOptions] = useState({});
@@ -131,6 +114,7 @@ export default function List({ footerLinks, userTokenInfo }) {
   const [academicYear, setAcademicYear] = useState();
   const [isDrawerOpen, setIsDrawerOpen] = useState(false);
   const [urlFilterApply, setUrlFilterApply] = useState(false);
+  const navigate = useNavigate();
 
   const handleOpenButtonClick = () => {
     setIsDrawerOpen((prevState) => !prevState);
@@ -150,13 +134,6 @@ export default function List({ footerLinks, userTokenInfo }) {
       setEnumOptions(data?.data ? data?.data : {});
 
       let newSchema = getOptions(schemat, {
-        key: "status",
-        arr: data?.data?.FACILITATOR_STATUS,
-        title: "title",
-        value: "value",
-      });
-
-      newSchema = getOptions(newSchema, {
         key: "district",
         arr: getDistricts?.districts,
         title: "district_name",
@@ -202,12 +179,13 @@ export default function List({ footerLinks, userTokenInfo }) {
     const fetchFilteredData = async () => {
       if (urlFilterApply) {
         setTableLoading(true);
-        const result = await facilitatorRegistryService.filter({
+        const result = await cohortService.Pclist({
           ...filter,
           limit: filter?.limit || 10,
         });
+        console.log({ result });
+        setData(result.data);
 
-        setData(result.data?.data);
         setPaginationTotalRows(result?.data?.totalCount || 0);
 
         setTableLoading(false);
@@ -232,7 +210,7 @@ export default function List({ footerLinks, userTokenInfo }) {
   );
 
   useEffect(() => {
-    const arr = ["district", "block", "status"];
+    const arr = ["district", "block"];
     const data = urlData(arr);
     if (Object.keys(data).find((e) => arr.includes(e))?.length) setFilter(data);
     setUrlFilterApply(true);
@@ -240,12 +218,8 @@ export default function List({ footerLinks, userTokenInfo }) {
 
   const onChange = useCallback(
     async (data) => {
-      const {
-        district: newDistrict,
-        block: newBlock,
-        status: newStatus,
-      } = data?.formData || {};
-      const { district, block, status, ...remainData } = filter || {};
+      const { district: newDistrict, block: newBlock } = data?.formData || {};
+      const { district, block, ...remainData } = filter || {};
       const finalFilter = {
         ...remainData,
         ...(newDistrict && newDistrict?.length > 0
@@ -254,7 +228,7 @@ export default function List({ footerLinks, userTokenInfo }) {
               ...(newBlock?.length > 0 ? { block: newBlock } : {}),
             }
           : {}),
-        ...(newStatus && newStatus?.length > 0 ? { status: newStatus } : {}),
+        // ...(newStatus && newStatus?.length > 0 ? { status: newStatus } : {}),
       };
       setFilterObject(finalFilter);
     },
@@ -265,10 +239,6 @@ export default function List({ footerLinks, userTokenInfo }) {
     setFilter({});
     setFilterObject({});
   }, [setFilterObject]);
-
-  const exportPrerakCSV = async () => {
-    await facilitatorRegistryService.exportFacilitatorsCsv(filter);
-  };
 
   const handleSearch = useCallback(
     (e) => {
@@ -324,22 +294,7 @@ export default function List({ footerLinks, userTokenInfo }) {
 
         <HStack height={"5vh"} space={2}>
           <AdminTypo.Secondarybutton
-            onPress={() => {
-              exportPrerakCSV();
-            }}
-            rightIcon={
-              <IconByName
-                color="#084B82"
-                _icon={{}}
-                size="15px"
-                name="ShareLineIcon"
-              />
-            }
-          >
-            {t("EXPORT")}
-          </AdminTypo.Secondarybutton>
-          <AdminTypo.Secondarybutton
-            // onPress={() => setModal(true)}   nevigate to form
+            onPress={() => navigate("/admin/addpcuser")}
             rightIcon={
               <IconByName
                 color="#084B82"
@@ -351,49 +306,6 @@ export default function List({ footerLinks, userTokenInfo }) {
           >
             {t("ADD_PC")}
           </AdminTypo.Secondarybutton>
-
-          {/* <Modal
-            isOpen={modal}
-            onClose={() => setModal(false)}
-            safeAreaTop={true}
-            size="xl"
-          >
-            <Modal.Content>
-              <Modal.CloseButton />
-              <Modal.Header p="5" borderBottomWidth="0">
-                <AdminTypo.H3 textAlign="center" color="textMaroonColor.600">
-                  {t("SEND_AN_INVITE")}
-                </AdminTypo.H3>
-              </Modal.Header>
-              <Modal.Body p="5" pb="10">
-                <VStack space="5">
-                  <HStack
-                    space="5"
-                    borderBottomWidth={1}
-                    borderBottomColor="gray.300"
-                    pb="5"
-                  >
-                    <AdminTypo.H4> {t("INVITATION_LINK")}</AdminTypo.H4>
-                    <Clipboard
-                      text={`${process.env.REACT_APP_BASE_URL}/facilitator-self-onboarding?org_id=${userTokenInfo?.authUser?.program_users?.[0]?.organisation_id}&cohort_id=${academicYear?.academic_year_id}&program_id=${program?.program_id}`}
-                    >
-                      <HStack space="3">
-                        <IconByName
-                          name="FileCopyLineIcon"
-                          isDisabled
-                          rounded="full"
-                          color="blue.300"
-                        />
-                        <AdminTypo.H3 color="blue.300">
-                          {t("CLICK_HERE_TO_COPY_THE_LINK")}
-                        </AdminTypo.H3>
-                      </HStack>
-                    </Clipboard>
-                  </HStack>
-                </VStack>
-              </Modal.Body>
-            </Modal.Content>
-          </Modal> */}
         </HStack>
       </HStack>
       <HStack ml="-1">
@@ -475,7 +387,7 @@ export default function List({ footerLinks, userTokenInfo }) {
           rounded={"xs"}
           height={"50px"}
           bg={
-            filter?.district || filter?.state || filter?.block || filter?.status
+            filter?.district || filter?.state || filter?.block
               ? "textRed.400"
               : "#E0E0E0"
           }
@@ -485,10 +397,7 @@ export default function List({ footerLinks, userTokenInfo }) {
           <IconByName
             name={isDrawerOpen ? "ArrowLeftSLineIcon" : "FilterLineIcon"}
             color={
-              filter?.state ||
-              filter?.district ||
-              filter?.block ||
-              filter?.status
+              filter?.state || filter?.district || filter?.block
                 ? "white"
                 : "black"
             }
@@ -534,8 +443,8 @@ const pagination = [10, 15, 25, 50, 100];
 
 const columns = (t, navigate) => [
   {
-    name: t("PRERAK_ID"),
-    selector: (row) => row?.id,
+    name: t("USER_ID"),
+    selector: (row) => row?.user_id,
     sortable: true,
     attr: "id",
     wrap: true,
@@ -543,29 +452,13 @@ const columns = (t, navigate) => [
     compact: true,
   },
   {
-    name: t("NAME"),
+    name: t("COORDINATORS_NAME"),
     selector: (row) => (
       <HStack display="inline-block" width={"100%"}>
         <Pressable
           style={{ flexDirection: "row", justifyContent: "space-between" }}
-          onPress={() => navigate(`/admin/facilitator/${row?.id}`)}
+          onPress={() => navigate(`/admin/facilitator/${row?.user_id}`)}
         >
-          {/* <HStack alignItems={"center"}> */}
-          {/* {row?.profile_photo_1?.name ? (
-                <ImageView
-                  urlObject={row?.profile_photo_1}
-                  alt="Alternate Text"
-                  width={"35px"}
-                  height={"35px"}
-                />
-              ) : (
-                <IconByName
-                  isDisabled
-                  name="AccountCircleLineIcon"
-                  color="gray.300"
-                  _icon={{ size: "35" }}
-                />
-              )} */}
           <AdminTypo.H6 bold word-wrap="break-word">
             {`${row?.first_name} ${row?.last_name || ""}`}
           </AdminTypo.H6>
@@ -580,102 +473,38 @@ const columns = (t, navigate) => [
     compact: true,
   },
   {
+    name: t("CONTACT"),
+    selector: (row) => row?.mobile || "-",
+    compact: true,
+  },
+  {
+    name: t("EMAIL_ID"),
+    selector: (row) => row?.email_id || "-",
+    compact: true,
+  },
+  {
+    name: t("STATE"),
+    selector: (row) => row?.state || "-",
+    compact: true,
+  },
+  {
     name: t("DISTRICT"),
     selector: (row) => row?.district || "-",
     compact: true,
   },
-  {
-    name: t("BLOCK"),
-    selector: (row) => row?.block || "-",
-    compact: true,
-  },
 
   {
-    name: t("MOBILE_NUMBER"),
-    selector: (row) => row?.mobile,
-    attr: "mobile",
-    wrap: true,
-    compact: true,
-  },
-  {
-    name: t("STATUS"),
-    selector: (row) => (
-      <Pressable onPress={() => navigate(`/admin/facilitator/${row?.id}`)}>
-        <ChipStatus py="0.5" px="1" status={row?.program_faciltators?.status} />
-      </Pressable>
-    ),
-    wrap: true,
-    attr: "status",
-    width: "150px",
-    compact: true,
-  },
-  {
-    name: (
-      <VStack display="inline-block" width={"100%"}>
-        {t("OKYC_VERIFICATION")}
-      </VStack>
-    ),
-    // wrap: true,
-    selector: (row) => {
-      return row?.aadhar_verified === "okyc_ip_verified"
-        ? t("OKYC_IP_VERIFIED")
-        : row?.aadhar_verified === "yes"
-        ? t("YES")
-        : t("NO");
-    },
-    compact: true,
-    minWidth: "50px",
-  },
-  {
-    minWidth: "140px",
     name: t("ACTION"),
     selector: (row) => (
-      <Button.Group
-        isAttached
-        divider={<div style={{ background: "#333", padding: "0.5px" }} />}
-        my="1"
-        h="6"
-        rounded={"full"}
-        shadow="BlueOutlineShadow"
-        borderWidth="1px"
+      <AdminTypo.Secondarybutton
+        my="3"
+        onPress={() => navigate(`/admin/pc/${row?.user_id}`)}
       >
-        <Button
-          background="white"
-          px="1.5"
-          _text={{
-            color: "blueText.400",
-            fontSize: "12px",
-            fontWeight: "700",
-          }}
-          onPress={() => {
-            navigate(`/admin/facilitator/${row?.id}`);
-          }}
-        >
-          {t("VIEW")}
-        </Button>
-        <Menu
-          w="190"
-          placement="bottom right"
-          trigger={(triggerProps) => dropDown(triggerProps, t)}
-        >
-          <Menu.Item
-            onPress={() => {
-              navigate(`/admin/facilitator/${row?.id}`);
-            }}
-          >
-            {t("VIEW")}
-          </Menu.Item>
-          <Menu.Item
-            onPress={() => {
-              navigate(`/admin/Certification/${row?.id}`);
-            }}
-          >
-            {t("RESULTS")}
-          </Menu.Item>
-        </Menu>
-      </Button.Group>
+        {t("ASSIGN_PRERAKS")}
+      </AdminTypo.Secondarybutton>
     ),
-    center: true,
+    sortable: true,
+    attr: "count",
   },
 ];
 
@@ -692,17 +521,9 @@ function Table({
   const [selectedData, setSelectedData] = useState();
   const navigate = useNavigate();
 
-  useEffect(() => {
-    const getData = async () => {
-      const result = await enumRegistryService.statuswiseCount();
-      setSelectedData(result);
-    };
-    getData();
-  }, []);
-
   const handleRowClick = useCallback(
     (row) => {
-      navigate(`/admin/pc/${row?.id}`);
+      navigate(`/admin/pc/${row?.user_id}`);
     },
     [navigate]
   );
@@ -711,44 +532,9 @@ function Table({
 
   return (
     <VStack>
-      <VStack p={2} pt="0">
-        <AdminTypo.H5 underline bold color="textMaroonColor.600">
-          {filter?.status === undefined || filter?.status?.length === 0 ? (
-            t("ALL") + `(${paginationTotalRows})`
-          ) : filter?.status?.[0] === "all" ? (
-            <AdminTypo.H4 bold>
-              {t("ALL") + `(${paginationTotalRows})`}
-            </AdminTypo.H4>
-          ) : (
-            filter?.status
-              ?.filter((item) => item)
-              .map(
-                (item) =>
-                  t(item).toLowerCase() +
-                  `(${
-                    selectedData
-                      ? selectedData?.find((e) => item === e.status)?.count
-                      : 0
-                  })`
-              )
-              .map((item) => item.charAt(0).toUpperCase() + item.slice(1))
-              .join(" , ")
-          )}
-        </AdminTypo.H5>
-      </VStack>
       <DataTable
         fixedHeader={true}
-        fixedHeaderScrollHeight={`${height - 160}px`}
-        customStyles={{
-          ...tableCustomStyles,
-          rows: {
-            style: {
-              minHeight: "20px", // override the row height
-              cursor: "pointer",
-            },
-          },
-          pagination: { style: { margin: "5px 0 5px 0" } },
-        }}
+        customStyles={tableCustomStyles}
         columns={columnsMemoized}
         data={data}
         persistTableHead
